backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
supabase==2.27.0
python-jose[cryptography]==3.3.0
google-generativeai==0.8.3
python-multipart==0.0.20
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
reportlab==4.2.0
